Renaming it from scheduling to spreadsheets.
Eventually we can put all the spreadsheet related utilities in here.
