+++ test/test_spreadsheets.js
@@ -1,6 +1,6 @@
 var assert = require('chai').assert;
 var moment = require("moment");
-var scheduling = require('../scheduling');
+var spreadsheets = require('../spreadsheets');
 
 var ISO_TUESDAY = 2;
 
@@ -8,12 +8,12 @@ describe('scheduling', function() {
     //--------------------------------------------------------------------------
     describe('#get_round_extrema()', function () {
         it(".isoWeekday() of the bounds should always return 2", function() {
-            var bounds = scheduling.get_round_extrema();
+            var bounds = spreadsheets.get_round_extrema();
             assert.equal(ISO_TUESDAY, bounds[0].isoWeekday());
             assert.equal(ISO_TUESDAY, bounds[1].isoWeekday());
         });
         it("The bounds containing 2016-04-15 are 2016-04-12T00:00:00 and 2016-04-19T:00:00:00 ", function() {
-            var bounds = scheduling.get_round_extrema({
+            var bounds = spreadsheets.get_round_extrema({
                 containing_date: moment.utc("2016-04-15")
             });
             start = bounds[0];
@@ -22,7 +22,7 @@ describe('scheduling', function() {
             assert.equal(end.format(), "2016-04-19T00:00:00+00:00")
         });
         it("The bounds containing 2016-04-15 but offset by an hour are 2016-04-11T23:00:00 and 2016-04-18T:23:00:00 ", function() {
-            var bounds = scheduling.get_round_extrema({
+            var bounds = spreadsheets.get_round_extrema({
                 containing_date: moment.utc("2016-04-15"),
                 offset_hours: 1
             });
@@ -46,7 +46,7 @@ describe('scheduling', function() {
                 offset_hours: 1
             };
             function test_parse_scheduling(string, expected)  {
-                var results = scheduling.parse_scheduling(string, options);
+                var results = spreadsheets.parse_scheduling(string, options);
                 assert.equal(results.date.format(), expected.date);
                 assert.equal(results.white, expected.white);
                 assert.equal(results.black, expected.black);
